// ev-backoffice/client/src/app/account/profile/edit-preview-profile/models/practice.enum.ts
export enum PracticeArea {

  BANKRUPTCY_AND_DEBT = 'BANKRUPTCY_AND_DEBT',
  BUSINESS = 'BUSINESS',
  CIVIL_RIGHTS = 'CIVIL_RIGHTS',
  CONSUMER_PROTECTION = 'CONSUMER_PROTECTION',
  CRIMINAL_DEFENSE = 'CRIMINAL_DEFENSE',
  EMPLOYMENT_AND_LABOR = 'EMPLOYMENT_AND_LABOR',
  ESTATE_PLANNING = 'ESTATE_PLANNING',
  FAMILY = 'FAMILY',
  GOVERNMENT = 'GOVERNMENT',
  IMMIGRATION = 'IMMIGRATION',
  INTELLECTUAL_PROPERTY = 'INTELLECTUAL_PROPERTY',
  LAWSUITS_AND_DISPUTES = 'LAWSUITS_AND_DISPUTES',
  PERSONAL_INJURY = 'PERSONAL_INJURY',
  REAL_ESTATE = 'REAL_ESTATE'

}